requirements.txt
Django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
djangorestframework-simplejwt
python-dotenv
